requirements.txt
asgiref==3.6.0
Django==3.2.18
django-allauth==0.54.0
django-crispy-forms==1.14.0
oauthlib==3.2.2
PyJWT==2.6.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests-oauthlib==1.3.1
sqlparse==0.4.4
